Move plunger to timer; Add dual-flip action
Instead of being in action vector, plunger now is called continuously on a timer.

The action that previously triggered the plunger is now used for a double-flip, in which the left and right sides are pressed simultaneously.

Also tweaked the ball loss penalty back to -0.3, to discourage ball draining and counteract ball end bonus score.